feat: add tool `send_message`
- add `tool/send_message.py`
- add minimum_dify_version
- fix sdk version
