+++ tools/outlook/requirements.txt
@@ -1,2 +1,2 @@
-dify_plugin>=0.3.0,<0.5.0
+dify_plugin>=0.4.2,<0.5.0
 requests>=2.31.0
