+++ tools/outlook/tools/send_draft.yaml
@@ -1,0 +1,40 @@
+identity:
+  name: send_draft
+  display_name: Send Draft
+  author: Dify
+  label:
+    en_US: Send Draft
+    zh_Hans: 发送草稿
+    pt_BR: Enviar Rascunho
+    ja_JP: 下書きを送信
+    zh_Hant: 發送草稿
+description:
+  human:
+    en_US: Send a draft email message through Outlook using Microsoft Graph API, requires a draft ID from the draft_message tool
+    zh_Hans: 通过Outlook使用Microsoft Graph API发送草稿电子邮件，需要从draft_message工具中获取草稿ID
+    pt_BR: Envia uma mensagem de e-mail rascunho através do Outlook usando a API do Microsoft Graph, requer um ID de rascunho do tool draft_message
+    ja_JP: Outlookを使用してMicrosoft Graph API経由で下書きメールを送信する。draft_messageツールから下書きIDが必要です。
+    zh_Hant: 通過Outlook使用Microsoft Graph API發送草稿電子郵件，需要從draft_message工具中獲取草稿ID
+  llm: Send a draft email message through Outlook using Microsoft Graph API, requires a draft ID from the draft_message tool
+
+parameters:
+  - name: draft_id
+    type: string
+    required: true
+    label:
+      en_US: Draft ID
+      zh_Hans: 草稿ID
+      pt_BR: ID do Rascunho
+      ja_JP: 下書きID
+      zh_Hant: 草稿ID
+    human_description:
+      en_US: The unique identifier of the draft message to send
+      zh_Hans: 要发送的草稿消息的唯一标识符
+      pt_BR: O identificador único do rascunho para enviar
+      ja_JP: 送信する下書きの一意の識別子
+      zh_Hant: 要發送的草稿訊息的唯一識別符
+    llm_description: The unique identifier of the draft message to send, this ID can be obtained from the draft_message tool
+    form: llm
+extra:
+  python:
+    source: tools/send_message.py
